Implementacion JDBC "Cuentas" #19 Refactorizacion de clases de implementacion
