+++ src/main/java/com/jlanzasg/novabank/model/Cuenta.java
@@ -3,9 +3,6 @@ package com.jlanzasg.novabank.model;
 import lombok.*;
 
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
-import java.util.Map;
 
 @ToString
 @Getter
