+++ src/main/java/com/jlanzasg/novabank/repository/Impl/MovimientoRepositoryImpl.java
@@ -1,14 +1,16 @@
-package com.jlanzasg.novabank.repository;
+package com.jlanzasg.novabank.repository.Impl;
 
 import com.jlanzasg.novabank.model.Movimiento;
+import com.jlanzasg.novabank.repository.MovimientoRepository;
 
+import java.sql.Connection;
 import java.time.LocalDateTime;
 import java.util.List;
 
 /**
  * The type Movimiento repository jdbc.
  */
-public class MovimientoRepositoryJdbc implements MovimientoRepository {
+public class MovimientoRepositoryImpl implements MovimientoRepository {
 
     @Override
     public Movimiento guardar(Movimiento movimiento) {
@@ -23,5 +25,10 @@ public class MovimientoRepositoryJdbc implements MovimientoRepository {
     @Override
     public List<Movimiento> buscarPorCuentaIdYFechas(Long cuentaId, LocalDateTime inicio, LocalDateTime fin) {
         return List.of();
+    }
+
+    @Override
+    public Movimiento guardar(Movimiento movimiento, Connection conn) {
+        return null;
     }
 }
